+++ src/builtin.rs
@@ -1,31 +1,36 @@
 mod array;
 mod class;
-pub mod errorobj;
+mod errorobj;
 mod fiber;
-pub mod file;
+mod file;
 mod float;
 mod hash;
 mod integer;
-pub mod math;
-pub mod method;
+mod math;
+mod method;
 mod module;
 mod object;
-pub mod process;
+mod process;
 mod procobj;
 mod range;
 mod regexp;
 mod string;
-pub mod structobj;
+mod structobj;
 pub use array::*;
 pub use class::*;
+pub use errorobj::*;
 pub use fiber::*;
+pub use file::*;
 pub use float::*;
 pub use hash::*;
 pub use integer::*;
+pub use math::*;
 pub use method::*;
 pub use module::*;
 pub use object::*;
+pub use process::*;
 pub use procobj::*;
 pub use range::*;
 pub use regexp::*;
 pub use string::*;
+pub use structobj::*;
